requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
bcrypt
pydantic-settings
python-multipart
httpx
apscheduler
anthropic
